requirements.txt
absl-py==2.0.0
affine==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
clearml==1.13.1
click-plugins==1.1.1
click==8.1.7
cligj==0.7.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
deformable-attention==0.0.18
descartes==1.1.0
dill==0.3.7
docker-pycreds==0.4.0
einops==0.7.0
executing==2.0.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
gitdb==4.0.11
gitpython==3.1.40
google-auth-oauthlib==1.0.0
google-auth==2.23.4
gpustat==1.1.1
grpcio==1.59.2
huggingface-hub==0.18.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipython==8.12.3
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsmin==3.0.1
jsonschema-specifications==2023.7.1
jsonschema==4.19.2
kiwisolver==1.4.5
kornia==0.7.0
lazy-loader==0.3
lightning-utilities==0.9.0
markdown==3.5.1
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.7.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.8.1.78
orderedmultidict==1.0.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathlib2==2.3.7.post1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.1
pip==23.3.2
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.40
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycocotools==2.0.7
pygments==2.16.1
pyjwt==2.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
pywavelets==1.4.1
pyyaml==6.0.1
qudida==0.0.4
rasterio==1.3.9
referencing==0.30.2
requests-oauthlib==1.3.1
requests==2.31.0
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.34.0
setproctitle==1.3.3
setuptools==67.7.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
stack-data==0.6.3
sympy==1.12
tensorboard-data-server==0.7.2
tensorboard==2.14.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.10
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wandb==0.16.2
wcwidth==0.2.9
werkzeug==3.0.1
wheel==0.40.0
x-transformers==1.24.5
yarl==1.9.2
zipp==3.17.0
